canvas drawing points, shapes
