+++ src/main/angular/lab-app/src/app/useful.ts
@@ -4,10 +4,10 @@ export function route(url: string, router: Router) {
   router.navigateByUrl(url).then(() => console.log('Navigated to ' + url))
 }
 
-export function canvasRelativeX(event: MouseEvent, canvasContainer: HTMLElement) {
+export function canvasRelativeX(event: MouseEvent, canvasContainer: HTMLElement): number {
   return event.pageX - canvasContainer.offsetLeft;
 }
 
-export function canvasRelativeY(event: MouseEvent, canvasContainer: HTMLElement) {
+export function canvasRelativeY(event: MouseEvent, canvasContainer: HTMLElement): number {
   return event.pageY - canvasContainer.offsetTop;
 }
